server/backend: split routes into api/routes/ + extract Pydantic schemas to models/
The pre-decomposition layout had route handlers, Pydantic schemas,
FastAPI dependencies, and the FastAPI app assembly all interleaved at
the top of the package. Pull each concern into its own subtree:

  api/
    deps.py            (renamed from cq_server/deps.py;
                        gains get_current_user previously in auth.py)
    routes/
      auth.py          (login route, extracted from auth.py)
      knowledge.py     (knowledge handlers, extracted from app.py)
      review.py        (renamed from cq_server/review.py)
      users.py         (renamed from cq_server/users.py)
  models/
    auth.py            (LoginRequest, LoginResponse)
    knowledge.py       (ProposeRequest, FlagRequest, StatsResponse)
    review.py          (ReviewItem, ReviewQueueResponse, ReviewDecisionResponse,
                        DailyCount, TrendsResponse, ReviewStatsResponse)
    users.py           (MeResponse, Message, CreateApiKeyRequest, ApiKeyPublic,
                        CreateApiKeyResponse, ApiKeysPublic)

Behaviour is unchanged: route bodies still take store directly via
get_store, the lifespan still owns a global SqliteStore. app.py shrinks
to lifespan + assembly + SPA fallback. auth.py shrinks to JWT and
password primitives.

Tests update their import paths (cq_server.deps -> cq_server.api.deps);
behaviour assertions are untouched.
