+++ server/backend/src/cq_server/api/deps.py
@@ -1,25 +1,33 @@
 """FastAPI dependencies shared across routers."""
 
-import hmac
+from __future__ import annotations
 
+import hmac
+import os
+
+import jwt
 from fastapi import BackgroundTasks, Depends, HTTPException, Request
 
-from .api_keys import decode_token, hash_secret
-from .store import Store
+from ..api_keys import decode_token, hash_secret
+from ..auth import verify_token
+from ..store import Store
 
 API_KEY_PEPPER_ENV = "CQ_API_KEY_PEPPER"  # pragma: allowlist secret
 
 
-def get_store(request: Request) -> Store:
-    """FastAPI dependency that returns the store from app state.
-
-    Args:
-        request: The incoming FastAPI request.
+def _get_jwt_secret() -> str:
+    """Return the JWT secret, failing if unset.
 
     Returns:
-        The Store instance attached to the application state.
+        The value of the CQ_JWT_SECRET environment variable.
+
+    Raises:
+        RuntimeError: If the environment variable is not set.
     """
-    return request.app.state.store
+    secret = os.environ.get("CQ_JWT_SECRET")
+    if not secret:
+        raise RuntimeError("CQ_JWT_SECRET environment variable is required")
+    return secret
 
 
 def get_api_key_pepper(request: Request) -> str:
@@ -32,6 +40,41 @@ def get_api_key_pepper(request: Request) -> str:
     if not pepper:
         raise HTTPException(status_code=500, detail="Server is misconfigured")
     return pepper
+
+
+def get_current_user(request: Request) -> str:
+    """FastAPI dependency that extracts and validates the JWT from the Authorization header.
+
+    Args:
+        request: The incoming FastAPI request.
+
+    Returns:
+        The username extracted from the validated token.
+
+    Raises:
+        HTTPException: With status 401 if the header is missing, malformed, or the token is invalid.
+    """
+    auth_header = request.headers.get("Authorization")
+    if not auth_header or not auth_header.startswith("Bearer "):
+        raise HTTPException(status_code=401, detail="Missing or invalid authorization header")
+    token = auth_header.removeprefix("Bearer ")
+    try:
+        payload = verify_token(token, secret=_get_jwt_secret())
+    except jwt.PyJWTError as exc:
+        raise HTTPException(status_code=401, detail="Invalid or expired token") from exc
+    return payload["sub"]
+
+
+def get_store(request: Request) -> Store:
+    """FastAPI dependency that returns the store from app state.
+
+    Args:
+        request: The incoming FastAPI request.
+
+    Returns:
+        The Store instance attached to the application state.
+    """
+    return request.app.state.store
 
 
 async def require_api_key(
