+++ server/backend/src/cq_server/api/routes/review.py
@@ -1,64 +1,17 @@
 """Review queue endpoints for the review API."""
 
-from cq.models import KnowledgeUnit
 from fastapi import APIRouter, Depends, HTTPException
-from pydantic import BaseModel
 
-from .auth import get_current_user
-from .deps import get_store
-from .store import Store
-
-
-class ReviewItem(BaseModel):
-    """A KU with its review metadata."""
-
-    knowledge_unit: KnowledgeUnit
-    status: str
-    reviewed_by: str | None
-    reviewed_at: str | None
-
-
-class ReviewQueueResponse(BaseModel):
-    """Paginated review queue response."""
-
-    items: list[ReviewItem]
-    total: int
-    offset: int
-    limit: int
-
-
-class ReviewDecisionResponse(BaseModel):
-    """Response after approving or rejecting a KU."""
-
-    unit_id: str
-    status: str
-    reviewed_by: str
-    reviewed_at: str
-
-
-class DailyCount(BaseModel):
-    """Daily proposal, approval, and rejection counts."""
-
-    date: str
-    proposed: int
-    approved: int
-    rejected: int
-
-
-class TrendsResponse(BaseModel):
-    """Trend data for the dashboard chart."""
-
-    daily: list[DailyCount]
-
-
-class ReviewStatsResponse(BaseModel):
-    """Dashboard metrics response."""
-
-    counts: dict[str, int]
-    domains: dict[str, int]
-    confidence_distribution: dict[str, int]
-    recent_activity: list[dict]
-    trends: TrendsResponse
+from ...models.review import (
+    DailyCount,
+    ReviewDecisionResponse,
+    ReviewItem,
+    ReviewQueueResponse,
+    ReviewStatsResponse,
+    TrendsResponse,
+)
+from ...store import Store
+from ..deps import get_current_user, get_store
 
 
 def _build_decision(unit_id: str, row: dict[str, str | None]) -> ReviewDecisionResponse:
