+++ server/backend/src/cq_server/auth.py
@@ -1,26 +1,22 @@
-"""Authentication: password hashing, JWT creation and validation."""
+"""Authentication primitives: password hashing and JWT creation/validation.
 
-import os
+Higher layers consume these helpers:
+
+- ``hash_password`` / ``verify_password`` are used by the login route and
+  by tests that seed users.
+- ``create_token`` issues a JWT after a successful login.
+- ``verify_token`` is consumed by the ``get_current_user`` FastAPI
+  dependency in ``api/deps.py``.
+
+Route handlers and FastAPI dependencies live elsewhere; this module is
+pure utility and has no FastAPI surface.
+"""
+
 from datetime import UTC, datetime, timedelta
 from typing import Any
 
 import bcrypt
 import jwt
-from fastapi import APIRouter, Depends, HTTPException, Request
-from pydantic import BaseModel
-
-from .deps import get_store
-from .store import Store
-
-
-def hash_password(password: str) -> str:
-    """Hash a password with bcrypt."""
-    return bcrypt.hashpw(password.encode(), bcrypt.gensalt()).decode()
-
-
-def verify_password(password: str, hashed: str) -> bool:
-    """Verify a password against its bcrypt hash."""
-    return bcrypt.checkpw(password.encode(), hashed.encode())
 
 
 def create_token(username: str, *, secret: str, ttl_hours: int = 24) -> str:
@@ -34,83 +30,16 @@ def create_token(username: str, *, secret: str, ttl_hours: int = 24) -> str:
     return jwt.encode(payload, secret, algorithm="HS256")
 
 
+def hash_password(password: str) -> str:
+    """Hash a password with bcrypt."""
+    return bcrypt.hashpw(password.encode(), bcrypt.gensalt()).decode()
+
+
+def verify_password(password: str, hashed: str) -> bool:
+    """Verify a password against its bcrypt hash."""
+    return bcrypt.checkpw(password.encode(), hashed.encode())
+
+
 def verify_token(token: str, *, secret: str) -> dict[str, Any]:
     """Verify and decode a JWT token."""
     return jwt.decode(token, secret, algorithms=["HS256"])
-
-
-class LoginRequest(BaseModel):
-    """Login request body."""
-
-    username: str
-    password: str
-
-
-class LoginResponse(BaseModel):
-    """Login response body."""
-
-    token: str
-    username: str
-
-
-def _get_jwt_secret() -> str:
-    """Return the JWT secret, failing if unset.
-
-    Returns:
-        The value of the CQ_JWT_SECRET environment variable.
-
-    Raises:
-        RuntimeError: If the environment variable is not set.
-    """
-    secret = os.environ.get("CQ_JWT_SECRET")
-    if not secret:
-        raise RuntimeError("CQ_JWT_SECRET environment variable is required")
-    return secret
-
-
-def get_current_user(request: Request) -> str:
-    """FastAPI dependency that extracts and validates the JWT from the Authorization header.
-
-    Args:
-        request: The incoming FastAPI request.
-
-    Returns:
-        The username extracted from the validated token.
-
-    Raises:
-        HTTPException: With status 401 if the header is missing, malformed, or the token is invalid.
-    """
-    auth_header = request.headers.get("Authorization")
-    if not auth_header or not auth_header.startswith("Bearer "):
-        raise HTTPException(status_code=401, detail="Missing or invalid authorization header")
-    token = auth_header.removeprefix("Bearer ")
-    secret = _get_jwt_secret()
-    try:
-        payload = verify_token(token, secret=secret)
-    except jwt.PyJWTError as exc:
-        raise HTTPException(status_code=401, detail="Invalid or expired token") from exc
-    return payload["sub"]
-
-
-router = APIRouter(prefix="/auth", tags=["auth"])
-
-
-@router.post("/login")
-async def login(request: LoginRequest, store: Store = Depends(get_store)) -> LoginResponse:
-    """Authenticate a user and return a JWT token.
-
-    Args:
-        request: Login credentials.
-        store: The store dependency.
-
-    Returns:
-        A LoginResponse with a signed JWT and the username.
-
-    Raises:
-        HTTPException: With status 401 if credentials are invalid.
-    """
-    user = await store.get_user(request.username)
-    if user is None or not verify_password(request.password, user["password_hash"]):
-        raise HTTPException(status_code=401, detail="Invalid username or password")
-    token = create_token(request.username, secret=_get_jwt_secret())
-    return LoginResponse(token=token, username=request.username)
