+++ server/backend/src/cq_server/models/auth.py
@@ -1,0 +1,17 @@
+"""Pydantic schemas for authentication routes."""
+
+from pydantic import BaseModel
+
+
+class LoginRequest(BaseModel):
+    """Login request body."""
+
+    username: str
+    password: str
+
+
+class LoginResponse(BaseModel):
+    """Login response body."""
+
+    token: str
+    username: str
