+++ server/backend/src/cq_server/models/knowledge.py
@@ -1,0 +1,27 @@
+"""Pydantic schemas for /knowledge routes."""
+
+from cq.models import Context, FlagReason, Insight
+from pydantic import BaseModel, Field
+
+
+class FlagRequest(BaseModel):
+    """Request body for flagging a knowledge unit."""
+
+    reason: FlagReason
+
+
+class ProposeRequest(BaseModel):
+    """Request body for proposing a new knowledge unit."""
+
+    domains: list[str] = Field(min_length=1)
+    insight: Insight
+    context: Context = Field(default_factory=Context)
+    created_by: str = ""
+
+
+class StatsResponse(BaseModel):
+    """Response body for store statistics."""
+
+    total_units: int
+    tiers: dict[str, int]
+    domains: dict[str, int]
